refactor: standardize folder structure
